Update parsing regex, act more deterministically.
The "ip" utility's output has changed a little bit. Updated regex is more flexible to this change, so it will work with the old "ip" utility output and the newly-formatted output.

Previously, the recommended routes were calculated and displayed. Now, they are forced to turn on. This is an unsafe change for someone who is modifying a machine remotely, but is remarkably useful for a machine with local access.

Troubleshooting text has been added.

The script is now at least roughly aware of how dumb it is.

It also now actually supports more than just deluge, but hopefully any application that needs namespaced VPN controls to prevent routing breakages.